Create ChaltuProfile.md
Marta Chaltu is a liver, lover, believer and giver.
She is an enthusiastic soul with a glow for life. An artistic mind with a love for numbers and science. 
A poetry lover & musical heart. 
She has overcome obstacles and believes in learning, healing and growing.
She is a bright mind from a young age with an active mind always hungry to solve problems and help others.